Move imgadm-list to it's own place
