--- main.go
@@ -1,9 +1,0 @@
-package main
-
-import (
-	"os"
-)
-
-func main() {
-	os.Exit(0)
-}
